create hook for home page
