+++ messages/messages.ts
@@ -1,8 +1,10 @@
 export const labels = {
     loading: 'Cargando...',
     search: 'Buscar',
+    pokemon: 'Pokemon',
 };
 
 export const messages = {
    scroll: 'Desplázate para cargar más',
+   selectPokemon: 'Selecciona un pokemon',
 };
